[DOC] 테스트 폴더 변경 및 회원 차단하기 테스트 추가
