+++ account/tests/__init__.py
@@ -1,0 +1,3 @@
+from account.tests.test_sign_up import *
+from account.tests.test_login import *
+from account.tests.test_blacklist import *
